- fixed jarvistelebot issue 38
