+++ err.go
@@ -81,4 +81,6 @@ var (
 	ErrNoConnOrInvalidConn = errors.New("no connection or invalid connection")
 	// ErrNoCtrlInfo - no ctrlinfo
 	ErrNoCtrlInfo = errors.New("no ctrlinfo")
+	// ErrCannotFindNodeWithAddr - can not find node with addr
+	ErrCannotFindNodeWithAddr = errors.New("can not find node with addr")
 )
